Make MAMGR specialization selection thru Makefile
The MAMGR is specialized for modem family. Until now, it was
configured using Android properties.
The configuration of MAMGR is now done using XML files which have
to be copied on the target.

This patch introduces a new Makefile to provide common modules
to be used on product on which MAMGR is used, handling XML
configuration files management.

Change-Id: I9666037de61daa5caddfd2e37b43bee19b3435e1
Tracked-On: GMINL-5039
Signed-off-by: Frederic Boisnard <frederic.boisnard@intel.com>
Reviewed-on: 343060
Reviewed-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-by: Esclafer, CatherineX <catherinex.esclafer@intel.com>
Tested-by: Esclafer, CatherineX <catherinex.esclafer@intel.com>