Put back "eth0" for dhcpcd service definition in x86 legacy mix-in.
This simple mix-in fix is for an issue that was introduced in
Release 77 when device encryption was enabled for MRD7 and T100
via GMIN-1382 and GMIN-1733.  Looking at the changes in the release,
it was noticed that the “eth0” was removed from hdcpcd service
entry in the check-in.  Putting the “eth0” back fixes the two
related bugs (GMIN-2102 and GMIN-2125).  This fix was checked on
T100 and MRD7 and was earlier code reviewed by Vikas Gupta
(who made the previous change).

Issue: GMIN-2102
Issue: GMIN-2125

Depends-Change-Id: I7b2c348abe67e8c965e351093a5d3c7e3e5607af
Depends-Change-Id: I2ec0667b27457830dbe2b98461d7ef4bc1a90fdc
Depends-Change-Id: Ideda8ef2f87605e77f9ba800787c19367ddfb588
Change-Id: Id88c3f038946ea1d1c6cbe2deee798837dea5178