Audio/Bluetooth: Remove all rules using 'BtEnable'
This criterion is set by calling setParameter on the HAL with the
"bluetooth_enabled" key.  However, this call does not exist anymore in the AOSP
and we can only rely on the fact that the policy selects a SCO device.

This criterion is now useless and all references to it in config files must go
away.

Issue: GMIN-2602

Change-Id: Ibfe0ea6facc8d27958f1f3b89bf3bde006d185f7
Signed-off-by: David Wagner <david.wagner@intel.com>
Reviewed-on: 225996
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Bohlin, Johan <johan.bohlin@intel.com>
Tested-by: Bohlin, Johan <johan.bohlin@intel.com>