power_profile: Add base power_profile.xml for BYT-T-CR devices
This patch adds the power_profile.xml for irda device. It is needed in order
to display the Battery Data on Settings->Battery.

The condition to display the Battery Data (in PowerUsageSummary.java) is that
the current consumed is higher than 10 mA:
getAveragePower(PowerProfile.POWER_SCREEN_FULL) < 10

This patch sets this value to 300 mA:
<item name="screen.full">300

Please note that all the values contained in this power_profile.xml file are
not exact. If exact values are needed then this file has to be modified.

Issue: IRDA-316
Change-Id: Ib99a44d919a4f59ad501cac84f6fc35a53ca8773
Signed-off-by: Ivan Gomez Castellanos <ivan.gomez.castellanos@intel.com>
Signed-off-by: Daniel Leung <daniel.leung@intel.com>
[dl: change commit msg title to mention BYT-T-CR devices.]
Reviewed-on: 245844
Reviewed-by: Jovanovic, Radivoje <radivoje.jovanovic@intel.com>