Telephony: Update apns-conf with intel changes.
This patch updates the apn-conf file with intel
modifications done in Kitkat.

The following patches are taken without any modification.

#/c/126098/
#/c/159078/
#/c/201849/
#/c/302684/

The following patches are not needed
#/c/149946/ as the data is already
available in the later file from google.

#/c/126097/ not needed as it
was subsequently removed.

#/c/149245/ is not needed as
it is a similar change of updating to the latest conf from google.

No such APN present as in below in the latest google apns-conf
file #/c/128037/
it has changed to a different name and the data needed is already
there.

No such APN as in change
#/c/131921/
#/c/214571/
is present in the latest apns-conf file from google

Change-Id: I134e31c296ed93c9dcca983690b15659e4d1343a
Tracked-On: IMINAN-12641
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Reviewed-on: 309741
Reviewed-on: 326733
Reviewed-by: Svenberg, Stefan <stefan.svenberg@intel.com>
Tested-by: Svenberg, Stefan <stefan.svenberg@intel.com>