common: Remove obsolete catalog configurations
The purpose of this patch is to replace V2.1 with V2 modem for
IMS variants and switch HSIC configuration to SSIC as it is now
the default IPC, by removing the former SSIC configurations.
Currently default configurations are incomplete and needs to be
updated to be fully align with SSIC requirements.

Change-Id: I12fc07df177f24d1c39ef60827e6d2144724fa14
Tracked-On: IMINAN-34409
Signed-off-by: Petru Muraru <petrux.muraru@intel.com>
Reviewed-on: 359564
Reviewed-on: 362885
Reviewed-by: Svenberg, Stefan <stefan.svenberg@intel.com>
Tested-by: Svenberg, Stefan <stefan.svenberg@intel.com>