Common: Clean-up the usage of skip_restoring_network_selection
The purpose of this patch is to clean-up the usage of
skip_restoring_network_selection overlay.
For flashless_tablet_voice configuration, overlay_voice is added on top
of overlay_data. There is no need to duplicate
skip_restoring_network_selection in overlay_voice as it is already
present in overlay_data

Change-Id: I949008ecae0f8645005134547a214c9ab059eb6b
Tracked-On: IMINAN-35491
Signed-off-by: Petru Muraru <petrux.muraru@intel.com>
Reviewed-on: 363297
Reviewed-on: 366454
Reviewed-by: Paut, Frederic <frederic.paut@intel.com>
Tested-by: Paut, Frederic <frederic.paut@intel.com>