fix build error related to mvconfig
Other parts of the build expect the mvconfig FLS to be staged in
$(FLASHFILES_DIR), but there was no explicit Makefile rule to put them
there. Instead we had a crude rule linked to droidcore to copy them all
at once. Get rid of this. Get rid of the "mvconfig" phony target that
depended on it, it wasn't even documented in build_info.

Change-Id: I0ea026cdf55a47a7a8f159cafc9306a53b4c83cb
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: 382979
Reviewed-by: Dattatraya Kulkarni, Anisha <anisha.dattatraya.kulkarni@intel.com>
Reviewed-by: Bergeron, Michael <michael.bergeron@intel.com>
Reviewed-by: Geltz, Brad <brad.geltz@intel.com>