sofia3gr, ota: read fwu_image.bin from provdata.zip
The fwu_image.bin file has moved from a loose file under RADIO in
the tfp into the RADIO/provdata.zip archive. The releasetools.py
script will only find the fwu_image.bin from provdata.zip in the
case that the build was made without named variants. If there are
variants then the releasetools.py script will emit the proper
fwu_image.bin steps in the edify script but leave the inclusion
of the fwu_image in the ota package to the ota_deployment_fixup
step.

Change-Id: I343b4707834556ededcacb35a3aaf37c27747976
Tracked-On: GMINL-9627
Signed-off-by: Scott D Phillips <scott.d.phillips@intel.com>
Reviewed-on: 395553