Remove forced reboot from SoFIA OTA extension
The original CTE code had this reboot forced during
the OTA process.  This was causing a number of issues,
and is now prohibited by SELinux.

Another change to the AOSP recovery code will check the
BCB status, and prevent it from being cleared if a
bootloader/FW update is staged via the BCB.

Based on CTE patch:

    Don't reboot when in firmware update

    Change-Id: Ib38898331d6f793b43b0ebd3c467b2cba27b11fb
    Signed-off-by: gli41 <genshen.li@intel.com>
    Tracked-On: CTEAN-2470

Change-Id: I5dbbb62376aa186bcb68091ad19ac46bdfe49161
Tracked-On: GMINL-11924
Signed-off-by: Brad Geltz <brad.geltz@intel.com>
Reviewed-on: 398121