[CWS_WIDI_UX] Do not scan in background when P2P connected/ing.
When no WiFi network is remembered, wpa_supplicant do not
automatically schedule periodic scan.
In that case, Android framework take the action to trig a scan
with the same interval of time than "suplicant scan interval".
When P2P is connected, this Android framework scan is not triggered
and supplicant scan interval is changed.
Set it to 3600s in order to minimize scan impact on P2P
connection, mainly for Miracast purpose.

Change-Id: Ic560d0e1bcc629c1fd7d2c7f376850946dad55e5
Tracked-On: OAM-3747
Reviewed-on: 403917
Signed-off-by: Jean Trivelly <jean.trivelly@intel.com>
Signed-off-by: Aymen Ben Alaya <aymenx.ben.alaya@intel.com>
Reviewed-on: 415321