[CWS_WIDI_UX] Adapter reconnection upon AP just connected on another frequency
Since some WiFi chip has some limitation to offer high video quality while used
in muli-roles and multi-channels, the best solution is to use the same channel as
the access point when running in multi-roles.

This solution is already in place when DUT is first connected to an access point
and then connected to a video adapter. However, in some case we can imagagine that
user first connects to video adapter and then connects to an access point. In
such situation, there is a high probability that adapter and access point
doesn't run on the same channel.

This patch detects when DUT is reaching an access point with a different
frequency than video adapter, a dialog box informs the user that the video link
will be closed and restored for better quality. The user has no other choice
than accept and the patch manages the disconnection, retrieves the adapter we
left and finally performs the reconnection.

Change-Id: Id6cc776ef14384265fcaaa2e7b881d42d2794881
Tracked-On: OAM-5528
Signed-off-by: Jean Trivelly <jean.trivelly@intel.com>
Reviewed-on: 403918
(cherry picked from commit 4bf35c6ba0a6d7d5ec0a71abb1ca4dbcdd2bcd90)
Reviewed-on: 425040