[AUDIO] Fix and add new PFW debug scripts
The current "parameter" and "vibrator" scripts are broken on user-debug
builds. This is due to the netstat command.
We need to add the -l option which displays listening server sockets.

This patch adds a "parameter-route" script in order to control
the route manager PFW instance.

This patch also renames the "parameter" script into "parameter-audio".
This patch also renames the "vibrator" script into "parameter-vibrator".

Change-Id: I01f2741032118a15315653b48302cf4b8f698f82
Tracked-On: OAM-1754
Signed-off-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-on: 407566