audio: Allows format, rate and channels to take dynamic values
This patch updates the route manager structure to allow giving
the supported rate, format and / or channel a specific value "dynamic"
in order to inform the audio HAL that these settings must be retrieved
dynamically.
It is typically the case of HDMI devices. Their capabilities depends on
connected tv screen and must be read (EDID information).
Note that the convention used is aligned with Android Audio Framework.

Change-Id: I48137c002637b63ea86614281ee249c4aef637c3
Tracked-On: OAM-7451
Signed-off-by: François Gaffie <francois.gaffie@intel.com>
Reviewed-on: 427841