Add smp_profiling configuration and remove useless code
The smp_profiling configuration is needed as Debug and Trace AREQ.
The code to support multi configurations already exists for sofia 3GR,
so this patch just moves it in a common part for 3GR, Garnet and LTE.
It also align the way to define the default configuration through
MV_CONFIG_DEFAULT_TYPE variable.

Finally, it cleans the fwu image management.
At some point, the SYSTEM_SIGNED_FLS_LIST variable contained all the
mv configurations and so does the fwu image which was problematic to
apply the OTA update. So the FWU_PACKAGE_BIN_LIST was modified to
contain only the default configuration.
But now that SYSTEM_SIGNED_FLS_LIST contains only the default
configuration, there is no need to modify FWU_PACKAGE_BIN_LIST.

At last, this patch enables the multi configuration support for Sofia
3GR Garnet which uses specific makefiles.

Change-Id: I24be57d9381d52a09f4f424a6a09529c7921793a
Signed-off-by: Gaelle Nassiet <gaellex.nassiet@intel.com>
Tracked-On: OAM-4814
Reviewed-on: 432828