[ASUS_3GR_M] Makefile changes for ASU_SF_3GR lunch target to compile the Modem as part of the Android full system build.
It is a customer requirement to compile the 3G Modem as part of the Android full system build.
These makefile changes for the ASU_SF_3GR lunch target are required to do this.
This change does not affect the normal 1A build, which uses a modem binary.
The compile flag BUILD_OS_AGNOSTIC_FROM_SRC this will get auto set with OC6 Manifest being used.

Change-Id: I92ba4d1dfb8367fbbbef3e2515e7019b30234e10
Signed-off-by: abhijeet.kadam@intel.com
Tracked-On: OAM-8673
Reviewed-on: 437538