Deprecated v2.30 of FlsTool
In favor of v2.36 in the intel-imc tools folder.  The current
work on the signing scripts depends on the new version.

The "FlsTool" in this folder is now a bash script that will print
an error message along with the path to the current version of FlsTool
in the intel-imc folder and exit without further operation.

Change-Id: I9f2887f9b8e19ff7ff2a446b24d24091005b9038
Tracked-On: OAM-18686
Signed-off-by: Brad Geltz <brad.geltz@intel.com>
Reviewed-on: 465717