Enabled per variant splash screen configurations
Needed for OEM customization work, this patch sets up the
easily maintainable structure for the splash screen.  New
cohos3gr variants will need to have their splash_screen.jpg
placed in a subfolder in splash_image/ that is named after
the variant in the device_mapping.py.

To differentiate the logic for r2 Sofia3GR targets we check
the size of SOFIA_PROVDATA_VARIANTS.

Change-Id: I1bf64893f60945422d434a9383fe831cd0eab26e
Tracked-On: OAM-7949
Signed-off-by: Brad Geltz <brad.geltz@intel.com>
Signed-off-by: adattatr <anisha.dattatraya.kulkarni@intel.com>
Reviewed-on: 466247