Create new variable to detect module using LOCAL_MODULE_TAGS
When using LOCAL_MODULE_TAGS the module will be built unconditionally
(not depending on a specific target).
Using PRODUCT_PACKAGES defined in the makefile of the target is
preferred (in device/intel/...).

A rule is set in vendor/intel/build/tasks/flashfiles.mk
to extract the list of modules that defines LOCAL_MODULE_TAGS.

Change-Id: I3dd25cb563f1a37526fc3d9d5628488a71b8a952
Tracked-On: OAM-19572
Signed-off-by: Benjamin Pujol <benjamin.pujol@intel.com>
Reviewed-on: 468175