Create mixins configuration for app compatibility enhancement feature
This feature is created for some cerner cases of
app compatibility issue. It's disabled by default.
To enable it, please add below lines to mixins.spec.
Currently, it aims to handle below issues:

1. All native libraries are put under assets directory.
Some apps put their libraries under the assets directory
in their APKs, instead of lib directory. It's not a
problem if the ABI of app libs is matched with platform
default ABI. But for apps with other ABIs, this feature
must be enabled to call native bridge for help.

Change-Id: I50efb90be6dfd55d3bd51d5548c20ce5bfa9edc0
Tracked-On: OAM-24201
Signed-off-by: jgu21 <jinghui.gu@intel.com>
Reviewed-on: 481127