audio: Update rt5645 routings paths when playing multiple devices
Ihf and headset/headphone devices share the same digital mixer path.
When playing through a single devices, the route is set accordingly.
However some audio cases (ringtones or alarms being played via Ihf
and headset), the path is closed by one of the devices making the
second device with no audio (the digital interface was stop).

The solution is to keep the common switches enable and mute gains
in order not to introduce clicks and pops during the device switch.

Change-Id: Id980a8074b3861c96230522e4e95f76962a437b5
Tracked-On: OAM-27687
Depends-Change-Id: I204ec71b7faf9ca7b949a967ca6575818356d667
Signed-off-by: Alejandro Ochoa <alejandro.ochoa@intel.com>
Reviewed-on: 488400
Reviewed-on: 492076