audio: Fixed music silence while taking photos
OutputDevice is not changing correctly when headphone/headset is
removed.

When the device boots, it configures its output device as
speaker for all the strategies, when a headset/headphone is connected,
it changes its output device to Speaker = NotSelected, and
Device2 = headset/headphone in enforced_audible strategy. When the
headset is disconnected, the output device changes to
speaker = selected, but Device2 keeps in the old value in the same
strategy. This bad state causes that PFW reports 2 devices
(Speaker|Headset), due to the camera app plays a sound using
enforced_audible strateby. AudioPolicy mutes the stream causing a short
music silence when the photo is taken.

Change-Id: I282ad1c033aad27296fe7c7381a12da37c938aab
Tracked-On: OAM-30953
Signed-off-by: Francisco Mendez <francisco.mendez@intel.com>
Reviewed-on: 496064
Reviewed-on: 502319