Add a StreamPath routing step to manage SmartX Stream operations
SmartXbar is hidden behind a PFW plugin.
SmartXbar will not only do the routing but also starting the sources and
sinks which results in opening audio devices. This step used to be done
by audio hal itself. We need to ensure audio control are done before to
be able to audio devices. This is the reason why this patch splits
the step path into StreamPath and Path.

Change-Id: I81db4b0fc6ea23fa49dfab9c326712a6a074ef58
Tracked-On: OAM-41495
Signed-off-by: François Gaffie <francois.gaffie@intel.com>
Reviewed-on: 554226